src/services/llm_interface/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via langchain-together
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.49.0
    # via langchain-anthropic
anyio==4.9.0
    # via
    #   anthropic
    #   httpx
    #   openai
    #   starlette
asyncio==3.4.3
    # via -r requirements.in
attrs==25.3.0
    # via aiohttp
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via -r requirements.in
distro==1.9.0
    # via
    #   anthropic
    #   openai
fastapi==0.115.12
    # via -r requirements.in
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.2
    # via huggingface-hub
graphviz==0.20.3
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   langchain-mistralai
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-mistralai
huggingface-hub==0.30.2
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jiter==0.9.0
    # via
    #   anthropic
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-anthropic==0.3.12
    # via -r requirements.in
langchain-core==0.3.54
    # via
    #   -r requirements.in
    #   langchain-anthropic
    #   langchain-deepseek
    #   langchain-mistralai
    #   langchain-openai
    #   langchain-together
langchain-deepseek==0.1.3
    # via -r requirements.in
langchain-mistralai==0.2.10
    # via -r requirements.in
langchain-openai==0.3.14
    # via
    #   -r requirements.in
    #   langchain-deepseek
    #   langchain-together
langchain-together==0.3.0
    # via -r requirements.in
langsmith==0.3.32
    # via langchain-core
loguru==0.7.3
    # via -r requirements.in
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
openai==1.75.0
    # via langchain-openai
orjson==3.10.16
    # via langsmith
packaging==24.2
    # via
    #   gunicorn
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   pytest
pluggy==1.5.0
    # via pytest
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.3
    # via
    #   -r requirements.in
    #   anthropic
    #   fastapi
    #   langchain-anthropic
    #   langchain-core
    #   langchain-mistralai
    #   langsmith
    #   openai
pydantic-core==2.33.1
    # via pydantic
pytest==8.3.5
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   langchain-core
redis==5.2.1
    # via -r requirements.in
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   huggingface-hub
    #   langchain-together
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
ruff==0.11.6
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
starlette==0.46.2
    # via fastapi
tenacity==9.1.2
    # via langchain-core
tiktoken==0.9.0
    # via langchain-openai
tokenizers==0.21.1
    # via langchain-mistralai
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typing-extensions==4.13.2
    # via
    #   anthropic
    #   fastapi
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via -r requirements.in
yarl==1.20.0
    # via aiohttp
zstandard==0.23.0
    # via langsmith
